+++ README.md
@@ -46,10 +46,10 @@
 
 
 ## 更新日志
-- v1.3.1
-  - 优化历史记录裁剪逻辑
-  - 优化token计算逻辑
-  - 修复edit消息的bug
+- v1.4.0
+  - 支持多平台部署
+  - 添加`/redo`指令，重新发送或者修改上一条提问
+  - 添加多语言支持，使用`LANGUAGE`环境变量设置语言，目前支持`zh-CN`，`zh-TW`和`en`。默认为`zh-CN`。
     
 其他更新日志见[CHANGELOG.md](./doc/CHANGELOG.md)
