+++ doc/CHANGELOG.md
@@ -1,4 +1,9 @@
 # 更新日志
+- v1.3.1
+    - 优化历史记录裁剪逻辑
+    - 优化token计算逻辑
+    - 修复edit消息的bug
+    
 - v1.3.0
     - 添加token使用统计指令`/usage`
     - 添加系统信息指令`/system`
