+++ src/message.js
@@ -1,9 +1,10 @@
 import {CONST, DATABASE, ENV} from './env.js';
 import {Context} from './context.js';
-import {sendChatActionToTelegramWithContext, sendMessageToTelegramWithContext} from './telegram.js';
+import {sendMessageToTelegramWithContext} from './telegram.js';
 import {requestCompletionsFromChatGPT} from './openai.js';
 import {handleCommandMessage} from './command.js';
-import {errorToString, tokensCounter} from './utils.js';
+import {errorToString} from './utils.js';
+
 // import {TelegramMessage, TelegramWebhookRequest} from './type.d.ts';
 
 
@@ -251,23 +252,13 @@ async function msgHandleRole(message, context) {
 async function msgChatWithOpenAI(message, context) {
   try {
     console.log('Ask:'+message.text||'');
-    const historyDisable = ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH <= 0;
     setTimeout(() => sendChatActionToTelegramWithContext(context)('typing').catch(console.error), 0);
-    const historyKey = context.SHARE_CONTEXT.chatHistoryKey;
-    const {real: history, original: original} = await loadHistory(historyKey, context);
-
-    const answer = await requestCompletionsFromChatGPT(message.text, history, context);
-    if (!historyDisable) {
-      original.push({role: 'user', content: message.text || '', cosplay: context.SHARE_CONTEXT.role || ''});
-      original.push({role: 'assistant', content: answer, cosplay: context.SHARE_CONTEXT.role || ''});
-      await DATABASE.put(historyKey, JSON.stringify(original)).catch(console.error);
-    }
+    const answer = await requestCompletionsFromChatGPT(message.text, context);
     return sendMessageToTelegramWithContext(context)(answer);
   } catch (e) {
     return sendMessageToTelegramWithContext(context)(`Error: ${e.message}`);
   }
 }
-
 
 /**
  * 根据类型对消息进一步处理
@@ -346,99 +337,6 @@ async function loadMessage(request, context) {
   } else {
     throw new Error('Invalid message');
   }
-}
-
-/**
- * 加载真实TG消息
- *
- * @param {string} key
- * @param {Context} context
- * @return {Promise<Object>}
- */
-async function loadHistory(key, context) {
-  const initMessage = {role: 'system', content: context.USER_CONFIG.SYSTEM_INIT_MESSAGE};
-  const historyDisable = ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH <= 0;
-
-  // 判断是否禁用历史记录
-  if (historyDisable) {
-    initMessage.role = ENV.SYSTEM_INIT_MESSAGE_ROLE;
-    return {real: [initMessage], original: [initMessage]};
-  }
-
-  // 加载历史记录
-  let history = [];
-  try {
-    history = JSON.parse(await DATABASE.get(key));
-  } catch (e) {
-    console.error(e);
-  }
-  if (!history || !Array.isArray(history)) {
-    history = [];
-  }
-
-
-  let original = JSON.parse(JSON.stringify(history));
-
-  // 按身份过滤
-  if (context.SHARE_CONTEXT.role) {
-    history = history.filter((chat) => context.SHARE_CONTEXT.role === chat.cosplay);
-  }
-
-  history.forEach((item)=>{
-    delete item.cosplay;
-  });
-
-  const counter = await tokensCounter();
-
-  const trimHistory = (list, initLength, maxLength, maxToken) => {
-    // 历史记录超出长度需要裁剪
-    if (list.length > maxLength) {
-      list = list.splice(list.length - maxLength);
-    }
-    // 处理token长度问题
-    let tokenLength = initLength;
-    for (let i = list.length - 1; i >= 0; i--) {
-      const historyItem = list[i];
-      let length = 0;
-      if (historyItem.content) {
-        length = counter(historyItem.content);
-      } else {
-        historyItem.content = '';
-      }
-      // 如果最大长度超过maxToken,裁剪history
-      tokenLength += length;
-      if (tokenLength > maxToken) {
-        list = list.splice(i + 1);
-        break;
-      }
-    }
-    return list;
-  };
-
-  // 裁剪
-  if (ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH > 0) {
-    const initLength = counter(initMessage.content);
-    const roleCount = Math.max(Object.keys(context.USER_DEFINE.ROLE).length, 1);
-    history = trimHistory(history, initLength, ENV.MAX_HISTORY_LENGTH, ENV.MAX_TOKEN_LENGTH);
-    original = trimHistory(original, initLength, ENV.MAX_HISTORY_LENGTH * roleCount, ENV.MAX_TOKEN_LENGTH * roleCount);
-  }
-
-  // 插入init
-  switch (history.length > 0 ? history[0].role : '') {
-    case 'assistant': // 第一条为机器人，替换成init
-    case 'system': // 第一条为system，用新的init替换
-      history[0] = initMessage;
-      break;
-    default:// 默认给第一条插入init
-      history.unshift(initMessage);
-  }
-
-  // 如果第一条是system,替换role为SYSTEM_INIT_MESSAGE_ROLE
-  if (ENV.SYSTEM_INIT_MESSAGE_ROLE !== 'system' && history.length > 0 && history[0].role === 'system') {
-    history[0].role = ENV.SYSTEM_INIT_MESSAGE_ROLE;
-  }
-
-  return {real: history, original: original};
 }
 
 /**
