+++ src/openai.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-unused-vars */
 import {Context} from './context.js';
 import {DATABASE, ENV} from './env.js';
+import {tokensCounter} from './utils.js';
 
 /**
  * 发送消息到ChatGPT
@@ -10,8 +11,9 @@ import {DATABASE, ENV} from './env.js';
  * @param {Context} context
  * @return {Promise<string>}
  */
-export async function requestCompletionsFromChatGPT(message, history, context) {
-  console.log(`requestCompletionsFromChatGPT: ${message}`);
+async function requestCompletionsFromOpenAI(message, history, context) {
+  console.log(`requestCompletionsFromOpenAI: ${message}`);
+  console.log(`history: ${JSON.stringify(history, null, 2)}`);
   const body = {
     model: ENV.CHAT_MODEL,
     ...context.USER_CONFIG.OPENAI_API_EXTRA_PARAMS,
@@ -64,6 +66,33 @@ export async function requestImageFromOpenAI(prompt) {
 }
 
 /**
+ *
+ * @param {string} text
+ * @param {Context} context
+ * @param {function} modifier
+ * @return {Promise<string>}
+ */
+export async function requestCompletionsFromChatGPT(text, context, modifier) {
+  const historyDisable = ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH <= 0;
+  const historyKey = context.SHARE_CONTEXT.chatHistoryKey;
+  let history = await loadHistory(historyKey, context);
+  if (modifier) {
+    const modifierData = modifier(history, text);
+    history = modifierData.history;
+    text = modifierData.text;
+  }
+  const {real: realHistory, original: originalHistory} = history;
+  const answer = await requestCompletionsFromOpenAI(text, realHistory, context);
+  if (!historyDisable) {
+    originalHistory.push({role: 'user', content: text || '', cosplay: context.SHARE_CONTEXT.role || ''});
+    originalHistory.push({role: 'assistant', content: answer, cosplay: context.SHARE_CONTEXT.role || ''});
+    await DATABASE.put(historyKey, JSON.stringify(originalHistory)).catch(console.error);
+  }
+  return answer;
+}
+
+
+/**
  * 更新当前机器人的用量统计
  * @param {object} usage
  * @param {Context} context
@@ -94,3 +123,97 @@ async function updateBotUsage(usage, context) {
 
   await DATABASE.put(context.SHARE_CONTEXT.usageKey, JSON.stringify(dbValue));
 }
+
+/**
+ * 加载历史TG消息
+ *
+ * @param {string} key
+ * @param {Context} context
+ * @return {Promise<Object>}
+ */
+async function loadHistory(key, context) {
+  const initMessage = {role: 'system', content: context.USER_CONFIG.SYSTEM_INIT_MESSAGE};
+  const historyDisable = ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH <= 0;
+
+  // 判断是否禁用历史记录
+  if (historyDisable) {
+    initMessage.role = ENV.SYSTEM_INIT_MESSAGE_ROLE;
+    return {real: [initMessage], original: [initMessage]};
+  }
+
+  // 加载历史记录
+  let history = [];
+  try {
+    history = JSON.parse(await DATABASE.get(key));
+  } catch (e) {
+    console.error(e);
+  }
+  if (!history || !Array.isArray(history)) {
+    history = [];
+  }
+
+
+  let original = JSON.parse(JSON.stringify(history));
+
+  // 按身份过滤
+  if (context.SHARE_CONTEXT.role) {
+    history = history.filter((chat) => context.SHARE_CONTEXT.role === chat.cosplay);
+  }
+
+  history.forEach((item) => {
+    delete item.cosplay;
+  });
+
+  const counter = await tokensCounter();
+
+  const trimHistory = (list, initLength, maxLength, maxToken) => {
+    // 历史记录超出长度需要裁剪
+    if (list.length > maxLength) {
+      list = list.splice(list.length - maxLength);
+    }
+    // 处理token长度问题
+    let tokenLength = initLength;
+    for (let i = list.length - 1; i >= 0; i--) {
+      const historyItem = list[i];
+      let length = 0;
+      if (historyItem.content) {
+        length = counter(historyItem.content);
+      } else {
+        historyItem.content = '';
+      }
+      // 如果最大长度超过maxToken,裁剪history
+      tokenLength += length;
+      if (tokenLength > maxToken) {
+        list = list.splice(i + 1);
+        break;
+      }
+    }
+    return list;
+  };
+
+  // 裁剪
+  if (ENV.AUTO_TRIM_HISTORY && ENV.MAX_HISTORY_LENGTH > 0) {
+    const initLength = counter(initMessage.content);
+    const roleCount = Math.max(Object.keys(context.USER_DEFINE.ROLE).length, 1);
+    history = trimHistory(history, initLength, ENV.MAX_HISTORY_LENGTH, ENV.MAX_TOKEN_LENGTH);
+    original = trimHistory(original, initLength, ENV.MAX_HISTORY_LENGTH * roleCount, ENV.MAX_TOKEN_LENGTH * roleCount);
+  }
+
+  // 插入init
+  switch (history.length > 0 ? history[0].role : '') {
+    case 'assistant': // 第一条为机器人，替换成init
+    case 'system': // 第一条为system，用新的init替换
+      history[0] = initMessage;
+      break;
+    default:// 默认给第一条插入init
+      history.unshift(initMessage);
+  }
+
+  // 如果第一条是system,替换role为SYSTEM_INIT_MESSAGE_ROLE
+  if (ENV.SYSTEM_INIT_MESSAGE_ROLE !== 'system' && history.length > 0 && history[0].role === 'system') {
+    history[0].role = ENV.SYSTEM_INIT_MESSAGE_ROLE;
+  }
+
+  return {real: history, original: original};
+}
+
